+++ example/peripheral/dma/gdma/main.c
@@ -16,31 +16,64 @@
  * LastEditTime: 2022-06-17 08:17:59
  * Description:  This file is for gdma example that running shell task and open scheduler
  *
- * Modify History:
- *  Ver   Who  Date   Changes
- * ----- ------  -------- --------------------------------------
+ * Modify History:
+ *  Ver      Who           Date         Changes
+ * -----    ------       --------      --------------------------------------
+ *  1.0    zhugengyu     2022/7/27     init commit
+ *  2.0    liqiaozhong   2023/11/10    synchronous update with standalone sdk
+ *  3.0    liqiaozhong   2024/4/22     add no letter shell mode, adapt to auto-test system
  */
 
+#include <stdio.h>
+
+#include "FreeRTOS.h"
+
+#include "sdkconfig.h"
+
+#ifdef CONFIG_USE_LETTER_SHELL
 #include "shell.h"
 #include "shell_port.h"
-#include <stdio.h>
+#else
+#include "task.h"
+#include "gdma_memcpy.h"
+
+#define GDMA_EXAMPLE_TASK_PRIORITY 2
+
+void GdmaExampleTaskEntry()
+{
+    /* example functions */
+    FFreeRTOSRunGdmaMemcpy();
+
+    /* end flag */
+    printf("[test_end]\r\n");
+    vTaskDelete(NULL);
+}
+#endif
 
 int main(void)
 {
-    BaseType_t ret;
-
-    ret = LSUserShellTask() ;
+    BaseType_t ret = pdPASS; /* 定义一个创建信息返回值，默认为 pdPASS */
+#ifdef CONFIG_USE_LETTER_SHELL
+    ret = LSUserShellTask();
+#else
+    /* used in no-letter-shell mode */
+    ret = xTaskCreate((TaskFunction_t)GdmaExampleTaskEntry,    /* 任务入口函数 */
+                      (const char *)"GdmaExampleTaskEntry",    /* 任务名字 */
+                      (uint16_t)4096,                          /* 任务栈大小 */
+                      NULL,                                    /* 任务入口函数参数 */
+                      (UBaseType_t)GDMA_EXAMPLE_TASK_PRIORITY, /* 任务优先级 */
+                      NULL);                                   /* 任务句柄 */
+#endif
     if (ret != pdPASS)
     {
         goto FAIL_EXIT;
     }
 
-    /* ret = FFreeRTOSRunGdmaMemcpy(); */
-
-    vTaskStartScheduler(); /* 启动任务，开启调度 */
+    /* 启动任务，开启调度 */
+    vTaskStartScheduler();
     while (1); /* 正常不会执行到这里 */
 
 FAIL_EXIT:
-    printf("failed 0x%x \r\n", ret);
+    printf("GDMA example failed in main.c, the ret value is 0x%x. \r\n", ret);
     return 0;
 }