+++ example/peripheral/pwm/inc/pwm_example.h
@@ -25,21 +25,25 @@
 
 #ifndef PWM_EXAMPLE_H
 #define PWM_EXAMPLE_H
+#include "FreeRTOS.h"
 
 #ifdef __cplusplus
 extern "C"
 {
 #endif
 
-#if defined(CONFIG_TARGET_E2000D)||defined(CONFIG_TARGET_E2000Q)
+/************************** Constant Definitions *****************************/
+#if defined(CONFIG_E2000Q_DEMO_BOARD)||defined(CONFIG_E2000D_DEMO_BOARD)
 #define PWM_TEST_ID FPWM6_ID
 #define PWM_TEST_CHANNEL     FPWM_CHANNEL_0
-#elif defined(CONFIG_TARGET_PHYTIUMPI)
-#define PWM_TEST_ID FPWM2_ID
-#define PWM_TEST_CHANNEL     FPWM_CHANNEL_1
+#elif defined(CONFIG_FIREFLY_DEMO_BOARD)
+#define PWM_TEST_ID FPWM4_ID
+#define PWM_TEST_CHANNEL     FPWM_CHANNEL_0
 #endif
 /* pwm test */
-BaseType_t FFreeRTOSPwmCreate(u32 id);
+BaseType_t FFreeRTOSPwmDeadBandTaskCreate();
+BaseType_t FFreeRTOSPwmSingleChannelTaskCreate();
+BaseType_t FFreeRTOSPwmDualChannelTaskCreate();
 
 #ifdef __cplusplus
 }